% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firstclass.R
\name{firstclass<-}
\alias{firstclass<-}
\title{Add a class in first place}
\usage{
firstclass(x) <- value
}
\arguments{
\item{x}{the object the new class is assigned to. Needs to be an object that can be copied.}

\item{value}{character string with the new class name.}
}
\value{
The same object x, now belonging (additionally) to the class given
as \code{"value"}.
}
\description{
Add a new class name to the classes of an R object, and make it
the first in the list of classes the objects belongs to.
}
\examples{
x <- data.frame(u = 1:2, t = 3:4)
class(x)
firstclass(x) <- "my.first.class"
class(x)
firstclass(x) <- "data.frame"
}
\author{
Ute Hahn,  \email{ute@imf.au.dk}
}

