\name{is.simplist}
\alias{is.simplist}
\title{Check whether an object is a simplist}
\usage{
  is.simplist(x)
}
\arguments{
  \item{x}{any \code{R} object}
}
\value{
  \code{TRUE} if \code{x} belongs to class
  \code{"simplist"}, otherwise \code{FALSE}.
}
\description{
  Checks if an object belongs to class \code{"simplist"}.
}
\seealso{
  \code{\link{simplist}} for a constructor of this class.
}

