% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listplot.R
\name{Internal functions}
\alias{Internal functions}
\alias{retagList}
\title{Internal functions}
\usage{
retagList(arglist, tagnames)
}
\description{
Utility functions
}
\keyword{internal}

